Omit key in output if required is false and value is none
In Django REST Framework, if a field is set to required=False and the value is None, the key will be omitted from the output.

Currently in serpy, if required is set to False and a value is null, the output will contain a null value for that key.

This commit changes the behaviour to match DRF. If a value is None and a key is set to required=False, the key will not appear in the output.

Includes an updated unit test to reflect this behaviour. A test case for type coercion on None values was also created to better reflect the code under test.