Allow missing keys on objects
Previously, a missing key on an object defined in the serializer would raise an attribute exception when serpy tried to access the attribute, even if the required=False flag was set.

In Django REST Framework, a missing key on a non-required field would result in that field not being reflected in the output.

This commit provides compatibility with DRF behaviour in serpy. Now, if an attribute is not present on the object to be serialized, and that field is not required, it will not be present in the reflected output.

This commit also includes improved unit tests for these cases, including tests for both serpy's Serializer and DictSerializer classes in the case of required values.