Fixed: Missing field does not cause KeyError
Previously, if an object was missing an attribute or key and it was not required it would raise an error in the serializer.

This commit changes the behaviour so that objects with non-required keys will be omitted from the output, while required keys will raise a TypeError.

This works for method fields as well. If a method is not required, it will skip serializing this field in the object. If it is required, but the return value is None it will also skip the field in the output. If it is required and the object does not contain the attribute, it will raise a TypeError stating that this field is required.